requirements.txt
transformers
faiss-gpu
numpy
pandas